fix performance report show, and add test for it
